refactored handleClick in card.jsx into smaller functions
